move network models to shared project
