+++ app/src/androidTest/java/ru/olegivo/afs/schedules/network/ScheduleNetworkSourceImplTest.kt
@@ -26,7 +26,7 @@ import ru.olegivo.afs.clubs.network.ClubsNetworkSourceImpl
 import ru.olegivo.afs.common.network.AuthorizedApiTest
 import ru.olegivo.afs.schedules.data.ScheduleNetworkSource
 import ru.olegivo.afs.schedules.network.models.DomainClub
-import ru.olegivo.afs.schedules.network.models.Schedules
+import ru.olegivo.afs.shared.network.models.Schedules
 
 class ScheduleNetworkSourceImplTest : AuthorizedApiTest() {
