+++ app/src/main/java/ru/olegivo/afs/schedules/data/ScheduleNetworkSource.kt
@@ -19,7 +19,7 @@ package ru.olegivo.afs.schedules.data
 
 import ru.olegivo.afs.schedules.data.models.DataSchedule
 import ru.olegivo.afs.schedules.domain.models.Slot
-import ru.olegivo.afs.schedules.network.models.Schedules
+import ru.olegivo.afs.shared.network.models.Schedules
 
 interface ScheduleNetworkSource {
