+++ shared/src/commonMain/kotlin/ru/olegivo/afs/shared/network/models/Schedules.kt
@@ -15,18 +15,19 @@
  * AFS.
  */
 
-package ru.olegivo.afs.schedules.network.models
+package ru.olegivo.afs.shared.network.models
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ActivityX(
-    val color: String?,
-    val id: Int,
-    val title: String,
-    val type: String,
-    val typeId: Int,
-    val url: String?,
-    val youtubePreviewUrl: String?,
-    val youtubeUrl: String?
+data class Schedules(
+    // TODO: later val club: Club,
+    val dateSince: String,
+    val dateTo: String,
+    val entryEnabled: Boolean,
+    val isNew: Boolean,
+    val needSlots: Boolean,
+    val next: String?,
+    val prev: String?,
+    val schedule: List<Schedule>
 )
