+++ shared/src/commonMain/kotlin/ru/olegivo/afs/shared/network/models/Trainer.kt
@@ -15,9 +15,22 @@
  * AFS.
  */
 
-package ru.olegivo.afs.schedules.network.models
+package ru.olegivo.afs.shared.network.models
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Slot(val id: Long, val slots: Int?)
+data class Trainer(
+    val clubs: List<Int>,
+    val facePhoto: String?,
+    val facebookLink: String,
+    val id: String,
+    val instagramLink: String,
+    val phone: String?,
+    val photo: String?,
+    val position: String,
+    val sortOrder: Int,
+    val title: String,
+    val url: String,
+    val vkLink: String
+)
